made changes to app to set up for microservicing based on specs and also blueprinted the registration process.
