+++ .gitignore
@@ -4,3 +4,4 @@ terraform/.terraform.lock.hcl
 venv/
 __pycache__/
 .DS_Store
+application/secretstuff.py
