+++ application/app.py
@@ -9,35 +9,42 @@ from flask import (
     render_template_string,
     session,
 )
-from flask_mail import Mail, Message
+from flask_mail import Mail
 from flask_mysqldb import MySQL
-from classes import RegisterForm
+# from classes import RegisterForm
 import QUERYDB as db
 from upload_test import upload_test_blueprint
 from serve_quiz import serve_quiz_blueprint
+from registration import registration_blueprint
 from passlib.hash import sha256_crypt
-from itsdangerous import URLSafeTimedSerializer
+# from itsdangerous import URLSafeTimedSerializer
 from functools import wraps
 from threading import Thread
-from random import randint
+# from random import randint
 import socket
-import secretstuff
-from emailverifier import Client
+# import secretstuff
+# from emailverifier import Client
+import os, config
 
 
 app = Flask(__name__, template_folder="templates", static_url_path="/static")
+app.config.from_object(os.environ.get('FLASK_ENV') or 'config.DevelopementConfig')
 app.register_blueprint(upload_test_blueprint)
 app.register_blueprint(serve_quiz_blueprint)
+app.register_blueprint(registration_blueprint)
 
-app.config.update(
-    DEBUG=True,
-    SECRET_KEY= 'testkey', #needed for cookie sessions. add to secrets file later
-    MAIL_SERVER='smtp.gmail.com',
-	MAIL_PORT=465,
-	MAIL_USE_SSL=True,
-	MAIL_USERNAME = secretstuff.emailusername,
-	MAIL_PASSWORD = secretstuff.emailpassword
-)
+
+
+# app.config.update(
+#     DEBUG=True,
+#     SECRET_KEY= 'testkey', #needed for cookie sessions. add to secrets file later
+#     MAIL_SERVER='smtp.gmail.com',
+# 	MAIL_PORT=465,
+# 	MAIL_USE_SSL=True,
+# 	MAIL_USERNAME = secretstuff.emailusername,
+# 	MAIL_PASSWORD = secretstuff.emailpassword
+# )
+
 
 
 @app.route("/")
@@ -77,42 +84,40 @@ def login():
 	return render_template('login.html')
 
 
-@app.route("/register", methods=["GET", "POST"])
-def register():
-    form = RegisterForm(request.form)
-    if request.method == "POST" and form.validate():
-        fullname = form.name.data
-        email = form.email.data
-        teachercode = form.teachercode.data
+# @app.route("/register", methods=["GET", "POST"])
+# def register():
+#     form = RegisterForm(request.form)
+#     if request.method == "POST" and form.validate():
+#         fullname = form.name.data
+#         email = form.email.data
+#         teachercode = form.teachercode.data
         
-        data = client.get(email) 
-        if str(data.smtp_check) == 'False': 
-            flash('Invalid email, please provide a valid email address','danger')
-            return render_template('register.html', form=form)
+#         data = client.get(email) 
+#         if str(data.smtp_check) == 'False': 
+#             flash('Invalid email, please provide a valid email address','danger')
+#             return render_template('register.html', form=form)
             
-        send_confirmation_email(email) 
-        user_id = randint(0,900000)
-        if teachercode == "KURATEACH2022":
-            user_id = randint(900000,999999)
-        print(user_id)
-        
-        username = form.username.data
-        password = sha256_crypt.encrypt(str(form.password.data))
-        cur = db.db.cursor()
-        cur.execute(
-            "INSERT INTO users(user_id, username,fullname,email, password,confirmed) values(%s,%s,%s,%s,%s,0)",
-            (user_id, username, fullname, email, password),
-        )
-        db.db.commit()
-        cur.close()
-        flash(
-            "Thanks for registering!  Please check your email to confirm your email address.",
-            "success",
-        )
-        return redirect(url_for("login"))
-        # change in login function to redirect to warning page
+#         send_confirmation_email(email) 
+#         user_id = randint(0,900000)
+#         if teachercode == "KURATEACH2022":
+#             user_id = randint(900000,999999)        
+#         username = form.username.data
+#         password = sha256_crypt.encrypt(str(form.password.data))
+#         cur = db.db.cursor()
+#         cur.execute(
+#             "INSERT INTO users(user_id, username,fullname,email, password,confirmed) values(%s,%s,%s,%s,%s,0)",
+#             (user_id, username, fullname, email, password),
+#         )
+#         db.db.commit()
+#         cur.close()
+#         flash(
+#             "Thanks for registering!  Please check your email to confirm your email address.",
+#             "success",
+#         )
+#         return redirect(url_for("login"))
+#         # change in login function to redirect to warning page
 
-    return render_template("register.html", form=form)
+#     return render_template("register.html", form=form)
 
 
 @app.route("/studDash")
@@ -433,88 +438,89 @@ def send_async_email(app, msg):
         mail.send(msg)
 
 
-htmlbody = """
-Your account on <b>The Best</b> Quiz App was successfully created.
-Please click the link below to confirm your email address and
-activate your account:
+# htmlbody = """
+# Your account on <b>The Best</b> Quiz App was successfully created.
+# Please click the link below to confirm your email address and
+# activate your account:
   
-<a href="{{ confirm_url }}">{{ confirm_url }}</a>
- <p>
---
-Questions? Comments? Email </p>
-"""
+# <a href="{{ confirm_url }}">{{ confirm_url }}</a>
+#  <p>
+# --
+# Questions? Comments? Email </p>
+# """
 
 
-client = Client("at_rFxZz7zEX8CO8V5IDBfzexOe2fW8b")
+# client = Client("at_rFxZz7zEX8CO8V5IDBfzexOe2fW8b")
 
 
-def send_email(recipients, html_body):
-    try:
-        msg = Message(
-            "Confirm Your Email Address",
-            sender="nycazn724@gmail.com",
-            recipients=recipients,
-        )
-        msg.html = html_body
-        send_async_email(app, msg)
-        # return 'Mail sent!'
-        return
-    except Exception as e:
-        # return(str(e))
-        return
+# def send_email(recipients, html_body):
+#     try:
+#         msg = Message(
+#             "Confirm Your Email Address",
+#             sender="nycazn724@gmail.com",
+#             recipients=recipients,
+#         )
+#         msg.html = html_body
+#         send_async_email(app, msg)
+#         # return 'Mail sent!'
+#         return
+#     except Exception as e:
+#         # return(str(e))
+#         return
 
 
-def send_confirmation_email(user_email):
-    confirm_serializer = URLSafeTimedSerializer(app.config["SECRET_KEY"])
-    ##https://stackoverflow.com/questions/34902378/where-do-i-get-a-secret-key-for-flask
+# def send_confirmation_email(user_email):
+#     confirm_serializer = URLSafeTimedSerializer(app.config["SECRET_KEY"])
+#     ##https://stackoverflow.com/questions/34902378/where-do-i-get-a-secret-key-for-flask
+#     print(confirm_serializer)
+#     confirm_url = url_for(
+#         "confirm_email",
+#         token=confirm_serializer.dumps(user_email, salt="email-confirmation-salt"),
+#         _external=True,
+#     )
+#     print(confirm_url)
+#     local_ip = "127.0.0.1"  # "get_local_ip() changed for testing
+#     x = ""
+#     if "localhost" in confirm_url:
+#         x = confirm_url.split("localhost:5000")
+#     else:
+#         x = confirm_url.split("127.0.0.1:5000")
+#     confirm_url = x[0] + local_ip + ":5000" + x[1]
+#     html = render_template_string(htmlbody, confirm_url=confirm_url)
 
-    confirm_url = url_for(
-        "confirm_email",
-        token=confirm_serializer.dumps(user_email, salt="email-confirmation-salt"),
-        _external=True,
-    )
-    local_ip = "127.0.0.1"  # "get_local_ip() changed for testing
-    x = ""
-    if "localhost" in confirm_url:
-        x = confirm_url.split("localhost:5000")
-    else:
-        x = confirm_url.split("127.0.0.1:5000")
-    confirm_url = x[0] + local_ip + ":5000" + x[1]
-    html = render_template_string(htmlbody, confirm_url=confirm_url)
-
-    send_email([user_email], html)
-
-
-"""This here salts the secret key and creates a token based off the user email added to secret key"""
+#     send_email([user_email], html)
 
 
-@app.route("/confirm/<token>")
-def confirm_email(token):
-    try:
-        confirm_serializer = URLSafeTimedSerializer(app.config["SECRET_KEY"])
-        email = confirm_serializer.loads(
-            token, salt="email-confirmation-salt", max_age=3600
-        )
-    except:
-        flash("The confirmation link is invalid or has expired.", "error")
-        return redirect(url_for("login"))
+# """This here salts the secret key and creates a token based off the user email added to secret key"""
 
-    cur = db.db.cursor()
-    results = cur.execute("SELECT * from users where email = %s", [email])
-    if results > 0:
-        data = cur.fetchone()
-        email_confirmed = data[5]
-        if email_confirmed:
-            flash("Account already confirmed. Please login.", "info")
-        else:
-            results = cur.execute(
-                "UPDATE users SET confirmed = 1 where email = %s", [email]
-            )
-            db.db.commit()
-            cur.close()
-            flash("Thank you for confirming your email address!", "success")
-            return redirect(url_for("login"))
-        return redirect(url_for("index"))
+
+# @app.route("/confirm/<token>")
+# def confirm_email(token):
+#     try:
+#         confirm_serializer = URLSafeTimedSerializer(app.config["SECRET_KEY"])
+#         email = confirm_serializer.loads(
+#             token, salt="email-confirmation-salt", max_age=3600
+#         )
+#     except:
+#         flash("The confirmation link is invalid or has expired.", "error")
+#         return redirect(url_for("login"))
+
+#     cur = db.db.cursor()
+#     results = cur.execute("SELECT * from users where email = %s", [email])
+#     if results > 0:
+#         data = cur.fetchone()
+#         email_confirmed = data[5]
+#         if email_confirmed:
+#             flash("Account already confirmed. Please login.", "info")
+#         else:
+#             results = cur.execute(
+#                 "UPDATE users SET confirmed = 1 where email = %s", [email]
+#             )
+#             db.db.commit()
+#             cur.close()
+#             flash("Thank you for confirming your email address!", "success")
+#             return redirect(url_for("login"))
+#         return redirect(url_for("index"))
 
 ##end email
