+++ application/secretstuff.py
@@ -1,2 +1,2 @@
-emailpassword = "elfohrsytbtyltht"
+emailpassword = "hypcodyuumzpkphg"
 emailusername = "nycazn724@gmail.com"
