+++ application/templates/index.html
@@ -75,8 +75,8 @@
 	<div class="grid12-6">
 		<div class="inner_box">
 			<p>Ready to get Started?</p>
-			<a class="btn btn-primary" href="register.html" style="font-size: 1.2em; background-color:#6a63dd">Sign Up</a>
-			<a class="btn btn-primary" href="login.html" style="font-size: 1.2em; background-color:#6a63dd">Login</a>
+			<a class="btn btn-primary" href="register" style="font-size: 1.2em; background-color:#6a63dd">Sign Up</a>
+			<a class="btn btn-primary" href="login" style="font-size: 1.2em; background-color:#6a63dd">Login</a>
 		</div>
