+++ application/templates/login.html
@@ -33,7 +33,7 @@
         <button type="submit" class="btn btn-primary">Submit</button>
         <div class="signup">
           <p>Don't have an account?
-            <a href="register.html" id="register">Click here</a>.
+            <a href="register" id="register">Click here</a>.
           </p>
         </div>
       </form>
