+++ application/testingfutureapp/__init__.py
@@ -1,0 +1,25 @@
+'''We are disregarding this for now.'''
+
+# from flask import Flask
+# # from flask_migrate import Migrate, MigrateCommand
+# # from flask_mail import Mail, Message
+# # from flask_sqlalchemy import SQLAlchemy
+# # from flask_script import Manager, Command, Shell
+# # from flask_login import LoginManager
+# import os, config
+
+# # create application instance
+# app = Flask(__name__)
+# app.config.from_object(os.environ.get('FLASK_ENV') or 'config.DevelopementConfig')
+
+# # initializes extensions
+# # db = SQLAlchemy(app)
+# # mail = Mail(app)
+# # migrate = Migrate(app, db)
+# # login_manager = LoginManager(app)
+# # login_manager.login_view = 'login'
+
+# # import views
+# # from . import views
+# # from . import forum_views
+# # from . import admin_views
